requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
av==10.0.0
Babel==2.11.0
backcall==0.2.0
base58==2.1.1
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
black==21.4b2
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.6
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython @ file:///home/conda/feedstock_root/build_artifacts/cython_1673054051202/work
cython-bbox==0.1.3
decorator==5.1.1
detectron2==0.6+cu102
dill==0.3.6
dlib==19.24.0
easydict==1.10
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.89.1
ffmpy==0.3.0
filelock==3.8.2
fire==0.5.0
Flask==2.2.2
flask-babel==3.0.0
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.3
fvcore==0.1.5.post20221220
gast==0.5.3
gdown==4.5.3
gevent==22.10.2
geventhttpclient==2.0.8
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.16.2
greenlet==2.0.1
grpcio==1.41.0
h11==0.14.0
h5py==3.7.0
httpcore==0.16.3
httpx==0.23.3
hydra-core==1.3.1
idna==3.4
imageio==2.23.0
imgaug==0.4.0
importlib-metadata==5.2.0
imutils==0.5.4
iopath==0.1.9
ipfs-api==0.2.3
ipfsapi==0.4.4
ipfshttpclient==0.4.13.2
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lap==0.4.0
libclang==14.0.6
linkify-it-py==1.0.3
lmdb==1.4.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
meson==1.0.0
meson-python==0.12.0
mpmath==1.2.1
multiaddr==0.0.9
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==0.4.3
nanoid==2.0.0
nats-py==2.2.0
nats-python==0.8.0
netaddr==0.8.0
networkx==2.8.8
ninja==1.11.1
numpy==1.22.0
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.12.0
opencv-contrib-python==4.7.0.68
opencv-python==4.5.5.64
openpyxl==3.0.10
opt-einsum==3.3.0
orjson==3.8.5
packaging==22.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.2
paddlepaddle==2.4.1
pandas==1.5.2
parameterized==0.8.1
parso==0.8.3
pathspec==0.11.0
pdf2docx==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
portalocker==2.6.0
premailer==3.10.0
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycryptodome==3.16.0
pydantic==1.10.4
pydot==1.4.2
pydub==0.25.1
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==1.7.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pyproject-metadata==0.7.1
pyrsistent==0.19.3
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-multipart==0.0.5
python-rapidjson==1.9
pytorchvideo==0.1.5
pytz==2022.7.1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==6.0
rapidfuzz==2.13.7
rarfile==4.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
shapely==2.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.22.0
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.1
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.64.1
traitlets==5.8.0
tritonclient==2.22.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
varint==1.0.2
visualdl==2.5.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
x2paddle==1.4.0
yacs==0.1.8
yapf==0.32.0
yarl==1.8.2
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
